% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC_theme_ggplot.R
\docType{data}
\name{colorlist_SMC}
\alias{colorlist_SMC}
\title{SMC-Theme f체r ggplot und ggplotly}
\format{
An object of class \code{list} of length 9.
}
\usage{
colorlist_SMC
}
\description{
Die Kernfunktion ist SMC_theme_ggplot(); mit dieser werden alle Theming-Parameter 
f체r ggplot gesetzt. Sie muss zu Beginn eines jeden R-/Quarto-/R-Markdown-Dokuments aufgerufen werden.
Beim initialen Aufrufen im Dokument muss ausgew채hlt werden, ob der Fokus des Dokuments auf Plotly oder ggplot liegt.
ggplot_SMC_theme(use_plotly = TRUE/FALSE)
}
\details{
Abh채ngig davon, ob der Schwerpunkt einer Aussendung auf ggplot-PNGs oder 
mit ggplotly erstellten interaktiven Abbildungen liegen soll, muss eine 
der folgenden Einstellungen im (dynamischen) Dokument genutzt werden: 
Plotly-Dokumente nutzen bitte: fig-width 9, fig-height 8, fig-dpi 200
ggplot-Dokumente nutzen bitte: fig-width 9, fig-height 7.25, fig-dpi 300
}
\keyword{datasets}
