% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC_theme_ggplot.R
\name{colors_SMC}
\alias{colors_SMC}
\title{colors_SMC
Mit dieser Funktion kann man sich eine einzelne oder mehrer Farben aus der SMC-Palette
in beliebiger Reihenfolge als Vektor oder String ausgeben lassen.}
\usage{
colors_SMC(ColorNames = NULL, rev = FALSE)
}
\arguments{
\item{ColorNames}{Ein String mit einem einzelenn Farbnamen oder ein Vektor mit mehreren Farbnamen.}

\item{rev}{boolean, if TRUE: umgekehrte Reihenfolge der Farben ausgeben}
}
\value{
Ein String mit einem einzelenn Farbcode oder ein Vektor mit mehreren unnamed Farbcodes.
}
\description{
colors_SMC
Mit dieser Funktion kann man sich eine einzelne oder mehrer Farben aus der SMC-Palette
in beliebiger Reihenfolge als Vektor oder String ausgeben lassen.
}
