% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_SMC.R
\docType{data}
\name{colors_SMC_named}
\alias{colors_SMC_named}
\title{Predefined color list for SMC}
\format{
An object of class \code{list} of length 9.
}
\usage{
colors_SMC_named
}
\description{
A named list containing color codes.
}
\keyword{datasets}
