% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_SMC.R
\docType{data}
\name{colors_SMC_unnamed}
\alias{colors_SMC_unnamed}
\title{Predefined color list for SMC}
\format{
An object of class \code{character} of length 9.
}
\usage{
colors_SMC_unnamed
}
\description{
An unnamed list containing color codes.
}
\keyword{datasets}
