% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_SMC.R
\name{create_base_map}
\alias{create_base_map}
\title{Helper functions to create map visualizations with leaflet
create_base_map}
\usage{
create_base_map(..., base_map_parameters = list())
}
\arguments{
\item{...}{Additional parameters to be passed to leafletOptions().}

\item{theme_params}{list, list of the customized parameters that can be obtained via \code{\link[=get_base_map_parameters]{get_base_map_parameters()}}.}
}
\value{
Leaflet map widget
}
\description{
Function for initializing a leaflet map with customized parameters.
}
\details{
Wrapper function for leaflet() that initializes an empty leaflet map with customized parameters:
leafletOptions, setView, setMaxBounds, and prepends a html-extra_style.
}
\examples{
create_base_map(base_map_parameters = get_base_map_parameters())
}
