% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_SMC.R
\name{create_custom_palette}
\alias{create_custom_palette}
\title{create_custom_palette}
\usage{
create_custom_palette(
  low_color,
  mid_color,
  high_color,
  start_point,
  mid_point,
  end_point,
  exp_factor = 1
)
}
\arguments{
\item{low_color}{character, color for low values}

\item{mid_color}{character, color for mid values}

\item{high_color}{character, color for high values}

\item{start_point}{numeric, start point of the color scale}

\item{mid_point}{numeric, midpoint of the color scale}

\item{end_point}{numeric, end point of the color scale}

\item{exp_factor}{numeric, factor for the exponential scaling (linear scaling if 1)}
}
\value{
Function that maps numeric values to colors
}
\description{
Function, that allows for a diverging color palette with a shifted midpoint.
In addition to the linear scaling, the function allows for an exponential scaling of the color ramp.
}
\examples{
create_custom_palette("blue", "white", "red", 0, 50, 100, exp_factor = 1)
}
