% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_download_SMC.R
\name{filename_slug}
\alias{filename_slug}
\title{filename_slug}
\usage{
filename_slug(x)
}
\arguments{
\item{x}{character vector.}
}
\value{
The slugified character vector.
}
\description{
Build a filename component: lower case, German umlauts transliterated
(ae, oe, ue, ss), everything non-alphanumeric collapsed to "-".
}
\examples{
filename_slug("Vereinigtes K\u00f6nigreich")
}
