% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_SMC.R
\name{get_base_map_parameters}
\alias{get_base_map_parameters}
\title{get_base_map_parameters}
\usage{
get_base_map_parameters()
}
\value{
List of default parameters for creating a leaflet map
}
\description{
Function that returns a list of default parameters for creating a leaflet map.
Default parameters are tweaked such that they are suitable for a map of Germany.
}
\examples{
get_base_map_parameters()
}
