% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_SMC.R
\name{get_choropleth_layer_parameters}
\alias{get_choropleth_layer_parameters}
\title{get_choropleth_layer_parameters}
\usage{
get_choropleth_layer_parameters()
}
\value{
List of default parameters for adding a leaflet choropleth layer
}
\description{
Function that returns a list of default parameters for a choropleth layer in a leaflet map.
}
\examples{
get_choropleth_layer_parameters()
}
