% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_SMC.R
\name{get_circle_marker_parameters}
\alias{get_circle_marker_parameters}
\title{map_add_circle_marker_layer}
\usage{
get_circle_marker_parameters()
}
\value{
List of default parameters for addCircleMarkers()
}
\description{
Function that generates default parameters for addCircleMarkers()
}
\examples{
get_circle_marker_parameters()
}
