% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_SMC.R
\name{get_param}
\alias{get_param}
\title{get_param}
\usage{
get_param(plot_parameters, param_name, default_value)
}
\arguments{
\item{plot_parameters}{list, list of parameters.}

\item{param_name}{character, name of the parameter to be retrieved.}

\item{default_value}{any, default value to be returned if the parameter is not present.}
}
\description{
Helper that allows to get a parameter from a list of parameters or a default value if the parameter is not present.
}
\examples{
get_param(list(margin_SMC = 10), "margin_SMC", 8)
}
