% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotly_SMC.R
\name{get_SMC_ggplotly_default_parameters}
\alias{get_SMC_ggplotly_default_parameters}
\title{get_SMC_ggplotly_default_parameters}
\usage{
get_SMC_ggplotly_default_parameters()
}
\value{
List of default parameters for ggplotly_SMC().
}
\description{
Function that returns the default parameters that are customized in ggplotly_SMC().
This list can be modified and passed to ggplotly_SMC() via the ggplotly_params argument.
}
\details{
Note: ggplot uses pt as units for font sizes, linewidths and point sizes, whereas plotly uses px.
To account for this, the function size_in_pt() is used to convert from px to pt.
However, a complete match is difficult to achieve due to the different fonts being used (here: PlusJakartaSans, plotly: embeds Circular from the SMC-Website).
}
\examples{
get_SMC_theme_ggplot_default_parameters()
}
