% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphQL_get_table.R
\name{GraphQL_get_table_string}
\alias{GraphQL_get_table_string}
\title{Get Data from GraphQL API (raw query string)}
\usage{
GraphQL_get_table_string(
  querystring,
  datenserver = "https://data.smclab.io/v1/graphql"
)
}
\arguments{
\item{querystring}{a complete GraphQL query (including operation name or anonymous)}

\item{datenserver}{character string URL of the GraphQL endpoint}
}
\value{
tibble with the requested data
}
\description{
Get Data from GraphQL API (raw query string)
}
