% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_helper.R
\name{image_helper}
\alias{image_helper}
\title{image_helper.R}
\usage{
image_helper(
  plot,
  filename,
  filepath,
  extra_html_tags = "width = \\"100\%\\" data-zoom=\\"1\\"",
  csv_opt = TRUE,
  plotly = FALSE,
  caption = "",
  save_svg = TRUE
)
}
\arguments{
\item{plot}{ggplot-object: The ggplot-object to be used for static images.}

\item{filename}{string: The name of the image file, without file extension.}

\item{filepath}{file.path: The path to the folder in which the image file should be saved. CSV must be placed manually, PNG and SVG are stored automatically.}

\item{extra_html_tags}{string: Between \'\' further HTML-Params can be added, e.g. \'width = \"100\%\" data-zoom=\"1\"\'}

\item{csv_opt}{boolean, FALSE: no link to a CSV-Download will be provided; default: TRUE.}

\item{plotly}{boolean: FALSE, a PNG of the provided ggplot-object will be displayed. TRUE: no figure is shown, plotly-figures must be displayed manually bforehand. Default: FALSE}

\item{caption}{string: If not empty: text for the caption, e.g. sources.}

\item{save_svg}{boolean: Generation of the corresponding svg. Default: TRUE. For very detailed graphics, you should refrain from creating a sv-graphic, since the file size can become too large.}
}
\value{
none
}
\description{
Function to create generic image captions in dynamic documents.
For this purpose, HTML code is created, which contains a link to the data needed to create this figure.
Also includes links to download the figure as a PNG or SVG.
}
\details{
The image_helper() function returns a list containing the file path, the title including links,
an "alt" text (consisting of the image name) and possibly other HTML tags.
}
\examples{
library(ggplot2)
library(dplyr)
tmp_ggplot <- ggplot(data = mpg, aes(x = displ, y = hwy))
tmp_ggplot \%>\% image_helper("MPG", file.path(tempdir()))
}
