% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_SMC.R
\name{map_add_choropleth_layer}
\alias{map_add_choropleth_layer}
\title{map_add_choropleth_layer}
\usage{
map_add_choropleth_layer(
  map,
  polygon_data,
  polygon_color_mapping,
  polygon_label,
  choropleth_layer_parameters,
  leaflet_group
)
}
\arguments{
\item{map}{Leaflet map widget}

\item{polygon_data}{SpatialPolygonsDataFrame, data for the choropleth layer}

\item{polygon_color_mapping}{character vector, one color per value in polygon_data}

\item{polygon_label}{character vector, labels for the polygons}

\item{choropleth_layer_parameters}{list, list of the customized parameters that can be obtained via \code{\link[=get_choropleth_layer_parameters]{get_choropleth_layer_parameters()}}.}

\item{leaflet_group}{character, name of the group to which the layer belongs}
}
\description{
Wrapper function for addPolygons() that adds a choropleth layer to a leaflet map.
}
