% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querystring_to_tabname_and_vec.R
\name{querystring_to_tabname_and_vec}
\alias{querystring_to_tabname_and_vec}
\title{Create Table- and Variable-names from a GraphQL Query}
\usage{
querystring_to_tabname_and_vec(querystring)
}
\arguments{
\item{querystring}{character string containing a valid simple GraphQL-query. Must not contain additional parameter or nested structures.}
}
\value{
invisible returns a list with tabname and variables. Cats string into console for esy copy and pasting.
}
\description{
Helper function to extraxt tablename and vector of variable names from an simple GraphQL query
}
\examples{
querystring_to_tabname_and_vec(
"query MyQuery {
  smc_story_meta {
  story_no
  title
  type
  url
  ressort
  publication_date
  }
}")

}
