% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC_theme_ggplot.R
\name{SMC_theme_ggplot}
\alias{SMC_theme_ggplot}
\title{ggplot_SMC_theme}
\usage{
SMC_theme_ggplot(use_plotly = FALSE)
}
\arguments{
\item{use_plotly}{boolean, TRUE: Aussendung ist auf plotly ausgelegt; FALSE: Aussendung ist auf ggplot-PNGs augelegt}
}
\value{
vector colorlist_SMC: alle SMC-Farben als Referenz
}
\description{
ggplot_SMC_theme
}
