% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC_theme_ggplot.R
\docType{data}
\name{colorlist_SMC}
\alias{colorlist_SMC}
\title{SMC-Theme for ggplot2 and plotly}
\format{
An object of class \code{list} of length 9.
}
\usage{
colorlist_SMC
}
\description{
Core functionality is SMC_theme_ggplot(). It sets all theme parameters for ggplot.
It must be called at the beginning of each R-/Quarto-/R-Markdown document.
}
\keyword{datasets}
