% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC_theme_ggplot.R
\docType{data}
\name{colorlist_SMC}
\alias{colorlist_SMC}
\title{SMC-Theme für ggplot und ggplotly}
\format{
An object of class \code{list} of length 9.
}
\usage{
colorlist_SMC
}
\description{
Die Kernfunktion ist SMC_theme_ggplot(); mit dieser werden alle Theming-Parameter 
für ggplot gesetzt. Sie muss zu Beginn eines jeden R-/Quarto-/R-Markdown-Dokuments aufgerufen werden.
}
\keyword{datasets}
