% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC_theme_ggplot.R
\name{colors_SMC}
\alias{colors_SMC}
\title{colors_SMC
This function returns one or more color codes from the SMC color palette
in arbitrary order as a vector or string (single color).}
\usage{
colors_SMC(ColorNames = NULL, rev = FALSE)
}
\arguments{
\item{ColorNames}{A string with a single color code or a vector with several colors.}

\item{rev}{boolean, if TRUE: reverse the return order of the colors.}
}
\value{
An unnamed vector with color codes or a single color code as a string.
}
\description{
colors_SMC
This function returns one or more color codes from the SMC color palette
in arbitrary order as a vector or string (single color).
}
