% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotly_SMC.R
\name{ggplotly_SMC}
\alias{ggplotly_SMC}
\title{ggplotly_SMC
Wrapper für ggplotly mit initialem Theming.}
\usage{
ggplotly_SMC(..., legende_unten = TRUE, mehrzeiliger_titel = FALSE)
}
\arguments{
\item{...}{Ein ggplot-Objekt, aus dem eine interaktive plotly-Abbildung erstellt werden soll. Außerdem: alle Parameter, die man üblicherweise an ggplotly übergeben würde, etwa:  ... = abbildung, tooltip = c("x", "y")}

\item{legende_unten}{boolean, FALSE: Legende rechts vom Plot; TRUE: Legende unter dem Plot}

\item{mehrzeiliger_titel}{boolean, TRUE: angepasstes Spacing zwischen Titel und Plot, falls der Titel mehrere Zeilen einnimmt}
}
\value{
plotly-object
}
\description{
ggplotly_SMC
Wrapper für ggplotly mit initialem Theming.
}
