% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotly_SMC.R
\name{ggplotly_SMC}
\alias{ggplotly_SMC}
\title{ggplotly_SMC
Eine Wrapperfunktion für ggplotly, die das Theming und einige Parameter anpasst, insbesondere
Schriftgröße, Abstände, Legendenpositionierung etc.}
\usage{
ggplotly_SMC(..., legende_unten = TRUE, mehrzeiliger_titel = FALSE)
}
\arguments{
\item{...}{Ein ggplot-Objekt, aus dem eine interaktive plotly-Abbildung erstellt werden soll. Außerdem: alle Parameter, die man üblicherweise an ggplotly übergeben würde, etwa:  ... = abbildung, tooltip = c("x", "y")}

\item{legende_unten}{boolean, FALSE: Legende rechts vom Plot; TRUE: Legende unter dem Plot}

\item{mehrzeiliger_titel}{boolean, TRUE: angepasstes Spacing zwischen Titel und Plot, falls der Titel mehrere Zeilen einnimmt}
}
\value{
plotly-object
}
\description{
ggplotly_SMC
Eine Wrapperfunktion für ggplotly, die das Theming und einige Parameter anpasst, insbesondere
Schriftgröße, Abstände, Legendenpositionierung etc.
}
\examples{
tmp_ggplot <- ggplot2::ggplot(data = ggplot2::mpg, aes(x = displ, y = hwy))
tmp_ggplotly <- tmp_ggplot \%>\% ggplotly_SMC()
tmp_ggplotly
tmp_ggplot \%>\% image_helper("MPG", file.path(tempdir()), plotly = TRUE)
}
