% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphQL_get_table.R
\name{GraphQL_get_table_vec}
\alias{GraphQL_get_table_vec}
\title{Get Data from GraphQL API (vector‑style)}
\usage{
GraphQL_get_table_vec(
  tabellenname,
  variablen,
  datenserver = "https://data.smclab.io/v1/graphql"
)
}
\arguments{
\item{tabellenname}{character string name of queried table}

\item{variablen}{character vector containing the desired variable names}

\item{datenserver}{character string URL of the GraphQL endpoint}
}
\value{
tibble with the requested data
}
\description{
Get Data from GraphQL API (vector‑style)
}
