% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_helper.R
\name{image_helper}
\alias{image_helper}
\title{image_helper.R}
\usage{
image_helper(
  plot,
  filename,
  filepath,
  extra_html_tags = "width = \\"100\%\\" data-zoom=\\"1\\"",
  csv_opt = TRUE,
  plotly = FALSE,
  caption = ""
)
}
\arguments{
\item{plot}{Das ggplot-Objekt was für die statischen Bilder genutzt werden soll.}

\item{filename}{string: Dateiname für die verlinkten Dateien (ohne Endungen).}

\item{filepath}{file.path: Dateipfad für die verlinkten Dateien. CSV muss manuell dort abgelegt werden, PNG und SVG werden von dieser Funktion dort abgelegt.}

\item{extra_html_tags}{string: Zwischen \'\' können weitere HTML-Parameter angegeben werden, etwa: \'width = \"100\%\" data-zoom=\"1\"\'}

\item{csv_opt}{boolean, FALSE: keine csv wird verlinkt; default: TRUE.}

\item{plotly}{boolean: FALSE, das oben übergegebene ggplot-Objekt wird angezeigt. TRUE: keine Abbildung wird angezeigt, plotly-Abbildungen müssen manuell vorher ausgegeben werden. Default: FALSE}

\item{caption}{string: Falls nicht leer: zusätzlicher Bild-Text, der unterhalb der Abbildung angezeigt werden kann. Etwa Datenquellen für Plotly-Abbildungen.}
}
\value{
none
}
\description{
Funktion, um generische Bildunterschriften in dynamischen Dokumenten für Abbildungen zu erstellen.
Hierfür wird HTML-Code erstellt, der einen Link auf die zur Erstellung dieser Abbildung benötigten Daten 
sowie Links zu der Abbildung als PNG und SVG enthält.
}
\details{
Die Funktion image_helper() gibt eine Liste zurück, die den Dateipfad, den Titel inklusive Verlinkungen, 
einen "alt" Text (bestehend aus dem Bildnamen) und ggf. weitere HTML-Tags enthält.
}
\examples{
library(ggplot2)
library(dplyr)
tmp_ggplot <- ggplot(data = mpg, aes(x = displ, y = hwy))
tmp_ggplot \%>\% image_helper("MPG", file.path(tempdir()))
}
