% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_helper.R
\name{image_helper}
\alias{image_helper}
\alias{image_helper_light}
\title{image_helper.R}
\usage{
image_helper(
  plot,
  filename,
  filepath,
  extra_html_tags = "width = \\"100\%\\" data-zoom=\\"1\\"",
  csv_opt = TRUE,
  plotly = FALSE,
  caption = "",
  caption_wrap_width = 100,
  save_svg = TRUE,
  fig_width = 9,
  fig_height = 8
)

image_helper_light(
  captions,
  filename = NULL,
  filepath = NULL,
  data = NULL,
  images = NULL,
  static_image = FALSE,
  static_image_parameters = ""
)
}
\arguments{
\item{plot}{ggplot-object: The ggplot-object to be used for static images.}

\item{filename}{string: The name of the image file, without file extension.}

\item{filepath}{file.path: The path to the folder in which the image file should be saved. (For image_helper(), CSV must be placed manually, PNG and SVG are stored automatically).}

\item{extra_html_tags}{string: Between \'\' further HTML-Params can be added, e.g. \'width = \"100\%\" data-zoom=\"1\"\'}

\item{csv_opt}{boolean, FALSE: no link to a CSV-Download will be provided; default: TRUE.}

\item{plotly}{boolean: FALSE, a PNG of the provided ggplot-object will be displayed. TRUE: no figure is shown, plotly-figures must be displayed manually bforehand. Default: FALSE}

\item{caption}{string: If not empty: text for the caption, e.g. sources.}

\item{caption_wrap_width}{integer: Width for wrapping the caption text. Default: 100.}

\item{save_svg}{boolean: Generation of the corresponding svg. Default: TRUE. For very detailed graphics, you should refrain from creating a sv-graphic, since the file size can become too large.}

\item{fig_width}{numeric: Width of the figure in inches. Default: 9.}

\item{fig_height}{numeric: Height of the figure in inches. Default: 8.}

\item{captions}{vector: A vector with all the captions for the plot.}

\item{data}{tibble: A dataframe with two columns 1) fileformat, and 2) filename_suffix. This dataframe is used to generate a link to each data set needed to generate the plot, usually there is one data set per plot, but sometimes there are more.}

\item{images}{tibble: A dataframe with two columns 1) fileformat, and 2) filename_suffix. This dataframe is used to generate links to files storing the plot in different image formats.}

\item{static_image}{boolean, FALSE: display static image above captions}

\item{static_image_parameters}{string: params to pass to img tag displaying static image}
}
\value{
Invisible: HTML code for image caption
}
\description{
Function to create generic image captions in dynamic documents.
For this purpose, HTML code is created, which contains a link to the data needed to create this figure.
Also includes links to download the figure in different formats.
}
\details{
The functions image_helper_light(), image_helper() return a list containing the file path, the title including links,
an "alt" text (consisting of the image name) and possibly other HTML tags.
}
\examples{
library(ggplot2)
library(dplyr)
tmp_ggplot <-
  ggplot(data = mpg, aes(x = displ, y = hwy)) +
  geom_point()
tmp_ggplot |> image_helper("MPG", file.path(tempdir()))
filename <- "a"
filepath <- tempdir()
data <- tibble(fileformat = c("csv", "json"), filename_suffix = c("", "extra_json_suffix"))
images <- tibble(fileformat = c("html"), filename_suffix = c(""))
captions = c("my_caption_1", "my_caption_2")
write(0, file = file.path(tempdir(), "a.csv"))
write(0, file = file.path(tempdir(), "a_extra_json_suffix.json"))
write(0, file = file.path(tempdir(), "a.html"))
image_helper_light(captions, filename, filepath, data, images)
}
