% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC_theme_ggplot.R
\name{SMC_theme_ggplot}
\alias{SMC_theme_ggplot}
\title{ggplot_SMC_theme}
\usage{
SMC_theme_ggplot()
}
\value{
Kein Rückgabewert, sondern setzt die ggplot2-Theme-Parameter.
}
\description{
Die Kernfunktion zum Erstellen von Quarto-Dokumenten im SMC-Theming: SMC_theme_ggplot().
Sie muss zu Beginn eines jeden Dokuments aufgerufen werden und stellt etwa die Schriftart (Aller),
die Schriftgröße, Paddings uvm. ein. Wirkt sich auch auf Plotly aus.
}
\examples{
SMC_theme_ggplot()
}
