% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC_theme_ggplot.R
\name{SMC_theme_ggplot}
\alias{SMC_theme_ggplot}
\title{ggplot_SMC_theme}
\usage{
SMC_theme_ggplot()
}
\value{
Invisibly returns the current theme.
}
\description{
Primary function for the creation of Quarto-Documents in the SMC-Theme: SMC_theme_ggplot().
It must be called at the beginning of each R-/Quarto-/R-Markdown document.
Sets font family, font size, padding, etc. Also affects Plotly.
}
\examples{
SMC_theme_ggplot()
}
