% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMC_theme_ggplot.R
\name{SMC_theme_ggplot}
\alias{SMC_theme_ggplot}
\title{SMC-Theme for ggplot2 and plotly}
\usage{
SMC_theme_ggplot(..., theme_params = list())
}
\value{
Invisibly returns the current theme.
}
\description{
Core functionality is SMC_theme_ggplot(). It sets all theme parameters for ggplot.
It must be called at the beginning of each R-/Quarto-/R-Markdown document.
}
\details{
ggplot_SMC_theme

Primary function for the creation of Quarto-Documents in the SMC-Theme: SMC_theme_ggplot().
It must be called at the beginning of each R-/Quarto-/R-Markdown document.
Sets font family, font size, padding, etc. Also affects Plotly.
}
\examples{
SMC_theme_ggplot()
}
